* swank-presentation-streams.lisp [sbcl]: Load the pretty-printer patch only at load time.  Add some trickery so that SBCL does not complain about the changed layout of the pretty-stream class.
* swank-presentation-streams.lisp (slime-stream-p): Using special
return values, indicate whether we are printing to the
REPL-results stream, or a dedicated stream.
(presentation-record): New slot "target".
(presentation-start, presentation-end): Use it (rather than the
global variable *use-dedicated-output-stream*) to decide whether
to use the bridge protocol or the :presentation-start/-end
protocol.  Also use it as the TARGET argument of
:presentation-start/-end messages.
(presenting-object-1): Use the new return values of
slime-stream-p.

* swank-presentation-streams.lisp: New file, moved here from
../present.lisp