Remove the slime-motd contrib
It hasn't worked out-of-the-box for a long time and it assumes the motd
file is a ChangeLog file which will be less useful when we remove
SLIME's ChangeLog files.

The idea was interesting, though; pull requests to revive this contrib
to grab content from sources like the NEWS file or the slime-tips blog
are welcome.